wip: kruskal part 2

+++ cpp/src/day08.hpp
@@ -4,5 +4,5 @@
 
 namespace day08 {
 int part1(const std::vector<std::string> &lines, int pairs);
-int part2(const std::vector<std::string> &lines);
+size_t part2(const std::vector<std::string> &lines);
 } // namespace day08
